updated ActiveTask to ActiveTaskTile to reflect the component tree
